requirements.txt

Pillow
pandas
plotly
requests
streamlit
